Added build. Closes #7.
This build closely follows the examples on http://docs.grafana.org/plugins/development/
