+++ .gitignore
@@ -14,3 +14,5 @@ hs_err_pid*
 .idea
 
 grafana-*
+
+node_modules
