+++ src/config-controller.js
@@ -1,0 +1,3 @@
+export class ChronixConfigController {}
+
+ChronixConfigController.templateUrl = 'partials/config.html';
