+++ src/partials/config.html
@@ -1,0 +1,2 @@
+<datasource-http-settings current="ctrl.current"></datasource-http-settings>
+
