+++ src/query-options-controller.js
@@ -1,0 +1,3 @@
+export class ChronixDBQueryOptionsCtrl {}
+
+ChronixDBQueryOptionsCtrl.templateUrl = 'partials/query.options.html';
